requirements.txt
# Core dependencies
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1

# Whisper and transformers
openai-whisper==20240930
transformers==4.36.2
datasets==2.16.1
accelerate==0.25.0
torch==2.1.2
torchaudio==2.1.2

# Audio processing and generation
pydub==0.25.1
audiomentations==0.33.0

# Optimization dependencies
resampy==0.4.2  # Fast resampling
gpustat==1.1.1  # GPU monitoring
GPUtil==1.4.0  # GPU monitoring for comparison script

# Evaluation metrics
jiwer==3.0.3
evaluate==0.4.1

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Development tools
jupyter==1.0.0
ipykernel==6.27.1
black==23.12.1
pytest==7.4.3 